requirements.txt
streamlit>=1.35.0
openai>=1.30.0
groq>=0.9.0
requests>=2.31.0
python-dotenv>=1.0.0
